backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
aiohttp==3.9.1
httpx==0.25.2

# Data Processing
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Blockchain APIs
web3==6.11.3
eth-account==0.9.0
eth-utils==2.3.1

# Task Queue & Scheduling
celery==5.3.4
redis==5.0.1
schedule==1.2.0

# Data Validation & Parsing
jsonschema==4.20.0
lxml==4.9.3
cssselect==1.2.0

# WebSocket Support
websockets==12.0
python-socketio==5.10.0

# Environment & Config
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Rate Limiting
ratelimit==2.2.1
tenacity==8.2.3

# Security
cryptography>=41.0.0
python-jose[cryptography]==3.3.0

# Machine Learning & AI
scikit-learn==1.3.2
scipy==1.11.4
joblib==1.3.2

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
sentencepiece==0.1.99
tokenizers==0.15.0
huggingface-hub==0.19.4

# Additional ML Libraries
imbalanced-learn==0.11.0
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.2
TextBlob==0.17.1

# Feature Engineering
feature-engine==1.6.2
category-encoders==2.6.3

# Model Deployment & Serving
onnx==1.15.0
onnxruntime==1.16.3

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
